Allow both the features to be used simultaneously
