+++ Cargo.toml
@@ -26,7 +26,7 @@ byteorder = "1.3.2"
 [features]
 tokio-traits = ["tokio"]
 futures-traits = ["futures-io"]
-#default = ["tokio-traits"]
+default = ["tokio-traits"]
 
 [dev-dependencies]
 tokio = "0.2.0-alpha.4"
